staging/src/main/java/com/hwtsllc/fixengine2022/fix42/tags/Tag266EnuAggregatedBook.java
/*
 * Copyright (c) 2022.  HW Tech Services, LLC
 *
 *   Licensed under the Apache License, Version 2.0 (the "License");
 *   you may not use this file except in compliance with the License.
 *   You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 *   Unless required by applicable law or agreed to in writing, software
 *   distributed under the License is distributed on an "AS IS" BASIS,
 *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *   See the License for the specific language governing permissions and
 *   limitations under the License.
 */

package com.hwtsllc.fixengine2022.fix42.tags;

import com.hwtsllc.fixengine2022.datatypes.FIX42;
import com.hwtsllc.fixengine2022.datatypes.FIX42Abstract;
import com.hwtsllc.fixengine2022.fix42.enums.Enum266AggregatedBook;
import com.hwtsllc.fixengine2022.interfaces.LogDataString;
import com.hwtsllc.fixengine2022.interfaces.LogValuePairString;
import com.hwtsllc.fixengine2022.interfaces.LogVerboseString;

/**
 *  266
 *  AggregatedBook
 *  Boolean
 *  <p>
 *  Specifies whether or not book entries should be aggregated. (Not specified) = broker option
 *  <p></p>
 *  Valid values:
 *  <p>    N - book entries should not be aggregated
 *  <p>    Y - book entries to be aggregated
 */
public class Tag266EnuAggregatedBook extends FIX42Abstract implements LogValuePairString, LogVerboseString, LogDataString {
    private final Enum266AggregatedBook dataValue;

    public final static Enum266AggregatedBook TESTA_ENU_AGGREGATED_BOOK
                    = Enum266AggregatedBook.NO ;
    public final static Enum266AggregatedBook TESTB_ENU_AGGREGATED_BOOK
                    = Enum266AggregatedBook.YES;

    public Tag266EnuAggregatedBook(Enum266AggregatedBook dataValue) {
        setFixType(FIX42.FIX266_ENU_AGGREGATED_BOOK);
        this.dataValue = dataValue;
    }

    /**
     * standard wrapper to retrieve the build a standard fix message for this tag
     */
    @Override
    public String toValuePairString() {
        return toFIXIDString()
                .concat("=")
                .concat(dataValue.toString());
    }
    /**
     * standard wrapper to format a detailed string describing this data field
     */
    @Override
    public String toVerboseString() {
        return super.toVerboseString()
                .concat("\n\tDataValue[")
                .concat(toString())
                .concat("]")
                .concat("\n\tValuePair[")
                .concat(toValuePairString())
                .concat("]")
                .concat("\n\tDataID[")
                .concat(toDataIDString())
                .concat("]")
                .concat("\n\tDataName[")
                .concat(toDataNameString())
                .concat("]")
                .concat("\n\tDataDescription[")
                .concat(toDataDescriptionString())
                .concat("]")
                ;
    }
    /**
     * wrapper to return the ID of the underlying Data
     */
    @Override
    public String toDataIDString() {
        return dataValue.toFIXIDString();
    }
    /**
     * wrapper to return the Name of the underlying Data
     */
    @Override
    public String toDataNameString() {
        return dataValue.toFIXNameString();
    }
    /**
     * wrapper to return the Description of the underlying Data
     */
    @Override
    public String toDataDescriptionString() {
        return dataValue.toEnumDescriptionString();
    }
    /**
     * standard wrapper to format a simple string describing the data
     */
    @Override
    public String toString() {
        return dataValue.toString();
    }

    /**
     *
     * @param args   no args used at this time
     */
    public static void main(String[] args) {
        Tag266EnuAggregatedBook tagData;

        tagData = new Tag266EnuAggregatedBook(TESTA_ENU_AGGREGATED_BOOK);
        System.out.println(tagData.toVerboseString());

        tagData = new Tag266EnuAggregatedBook(TESTB_ENU_AGGREGATED_BOOK);
        System.out.println(tagData.toVerboseString());

        // loop around the ENUM and display
        for (Enum266AggregatedBook oneEnum : Enum266AggregatedBook.values()) {
            System.out.println( new Tag266EnuAggregatedBook(oneEnum).toVerboseString() );
        }
    }
}
